refactor: decoupling required rp registration
```
$ acctests azurerm TestAccAzureRMEnsureRequiredResourceProvidersAreRegistered
=== RUN   TestAccAzureRMEnsureRequiredResourceProvidersAreRegistered
--- PASS: TestAccAzureRMEnsureRequiredResourceProvidersAreRegistered (6.53s)
PASS
ok  	github.com/terraform-providers/terraform-provider-azurerm/azurerm	8.404s
```
