+++ azurerm/required_resource_providers.go
@@ -1,5 +1,13 @@
 package azurerm
 
+import (
+	"context"
+	"log"
+
+	"github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2017-05-10/resources"
+	"github.com/terraform-providers/terraform-provider-azurerm/azurerm/helpers/resourceproviders"
+)
+
 // requiredResourceProviders returns all of the Resource Providers used by the AzureRM Provider
 // whilst all may not be used by every user - the intention is that we determine which should be
 // registered such that we can avoid obscure errors where Resource Providers aren't registered.
@@ -42,4 +50,21 @@ func requiredResourceProviders() map[string]struct{} {
 		"Microsoft.Sql":                 {},
 		"Microsoft.Storage":             {},
 	}
+}
+
+func ensureResourceProvidersAreRegistered(ctx context.Context, client resources.ProvidersClient, availableRPs []resources.Provider, requiredRPs map[string]struct{}) error {
+	log.Printf("[DEBUG] Determining which Resource Providers require Registration")
+	providersToRegister := resourceproviders.DetermineResourceProvidersRequiringRegistration(availableRPs, requiredRPs)
+
+	if len(providersToRegister) > 0 {
+		log.Printf("[DEBUG] Registering %d Resource Providers", len(providersToRegister))
+		err := resourceproviders.RegisterForSubscription(ctx, client, providersToRegister)
+		if err != nil {
+			return err
+		}
+	} else {
+		log.Printf("[DEBUG] All required Resource Providers are registered")
+	}
+
+	return nil
 }
